Code cleanup and finishing callouts
